+++ extension.ts
@@ -7,6 +7,7 @@ import { viewOthers } from './commands/viewOthers';
 import { importAndCompile } from './commands/compile';
 import { exportAll, exportExplorerItem } from './commands/export';
 import { xml2doc } from './commands/xml2doc';
+import { subclass } from './commands/subclass';
 
 import { ObjectScriptClassSymbolProvider } from './providers/ObjectScriptClassSymbolProvider';
 import { ObjectScriptRoutineSymbolProvider } from './providers/ObjectScriptRoutineSymbolProvider';
@@ -128,6 +129,7 @@ export async function activate(context: vscode.ExtensionContext): Promise<void> 
     vscode.commands.registerCommand('vscode-objectscript.compileWithFlags', () => importAndCompile(true)),
     vscode.commands.registerCommand('vscode-objectscript.export', exportAll),
     vscode.commands.registerCommand('vscode-objectscript.viewOthers', viewOthers),
+    vscode.commands.registerCommand('vscode-objectscript.subclass', subclass),
     vscode.commands.registerCommand('vscode-objectscript.touchBar.viewOthers', viewOthers),
     vscode.commands.registerCommand('vscode-objectscript.explorer.refresh', () => explorerProvider.refresh()),
     vscode.commands.registerCommand('vscode-objectscript.explorer.openClass', vscode.window.showTextDocument),