+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - Go to definition now goes to real file if such presented, or opens from the server
 - Basic syntax highlighting for CSP files, only as HTML
 - Added some snippets for class
+- View subclasses for current class
 
 ## [0.7.7]